Some cleanup related to the IDocumentProvider support
