+++ src/Swashbuckle.AspNetCore.SwaggerGen/Application/SwaggerGenOptionsExtensions.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Xml.XPath;
 using Microsoft.OpenApi.Models;
+using ApiExplorer = Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Swashbuckle.AspNetCore.SwaggerGen;
-using ApiDescriptionType = Microsoft.AspNetCore.Mvc.ApiExplorer.ApiDescription;
 
 namespace Microsoft.Extensions.DependencyInjection
 {
@@ -32,7 +32,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// </param>
         public static void DocInclusionPredicate(
             this SwaggerGenOptions swaggerGenOptions,
-            Func<string, ApiDescriptionType, bool> predicate)
+            Func<string, ApiExplorer.ApiDescription, bool> predicate)
         {
             swaggerGenOptions.SwaggerGeneratorOptions.DocInclusionPredicate = predicate;
         }
@@ -52,7 +52,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <param name="resolver"></param>
         public static void ResolveConflictingActions(
             this SwaggerGenOptions swaggerGenOptions,
-            Func<IEnumerable<ApiDescriptionType>, ApiDescriptionType> resolver)
+            Func<IEnumerable<ApiExplorer.ApiDescription>, ApiExplorer.ApiDescription> resolver)
         {
             swaggerGenOptions.SwaggerGeneratorOptions.ConflictingActionsResolver = resolver;
         }
@@ -62,7 +62,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// </summary>
         public static void CustomOperationIds(
             this SwaggerGenOptions swaggerGenOptions,
-            Func<ApiDescriptionType, string> operationIdSelector)
+            Func<ApiExplorer.ApiDescription, string> operationIdSelector)
         {
             swaggerGenOptions.SwaggerGeneratorOptions.OperationIdSelector = operationIdSelector;
         }
@@ -75,7 +75,7 @@ namespace Microsoft.Extensions.DependencyInjection
         [Obsolete("Deprecated: Use the overload that accepts a Func that returns a list of tags")]
         public static void TagActionsBy(
             this SwaggerGenOptions swaggerGenOptions,
-            Func<ApiDescriptionType, string> tagSelector)
+            Func<ApiExplorer.ApiDescription, string> tagSelector)
         {
             swaggerGenOptions.SwaggerGeneratorOptions.TagsSelector = (apiDesc) => new[] { tagSelector(apiDesc) };
         }
@@ -87,7 +87,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <param name="tagsSelector"></param>
         public static void TagActionsBy(
             this SwaggerGenOptions swaggerGenOptions,
-            Func<ApiDescriptionType, IList<string>> tagsSelector)
+            Func<ApiExplorer.ApiDescription, IList<string>> tagsSelector)
         {
             swaggerGenOptions.SwaggerGeneratorOptions.TagsSelector = tagsSelector;
         }
@@ -99,7 +99,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <param name="sortKeySelector"></param>
         public static void OrderActionsBy(
             this SwaggerGenOptions swaggerGenOptions,
-            Func<ApiDescriptionType, string> sortKeySelector)
+            Func<ApiExplorer.ApiDescription, string> sortKeySelector)
         {
             swaggerGenOptions.SwaggerGeneratorOptions.SortKeySelector = sortKeySelector;
         }
